requirements.txt
streamlit>=1.37.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
pandas>=2.2.2
plotly>=5.23.0
python-dateutil>=2.9.0
reportlab>=4.2.3
python-dotenv>=1.0.1
toml>=0.10.2
numpy>=1.26.4
werkzeug>=3.0.0
